+++ doc/changelog.rst
@@ -24,6 +24,7 @@ Localization
 Admin
 ~~~~~
 
+- Fix required permission issues for various modules
 - Make `model_url` context function and add permission check
 - Add permission check option to `get_model_url`
 - Add permission check to toolbar button classes